refactor: File*Service, JCF*Service 삭제 및 Basic*Service 추가
