+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicChannelService.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.service.file;
+package com.sprint.mission.discodeit.service.basic;
 
 import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
@@ -8,10 +8,10 @@ import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.UUID;
 
-public class FileChannelService implements ChannelService {
+public class BasicChannelService implements ChannelService {
     private final ChannelRepository channelRepository;
 
-    public FileChannelService(ChannelRepository channelRepository) {
+    public BasicChannelService(ChannelRepository channelRepository) {
         this.channelRepository = channelRepository;
     }
 
@@ -37,7 +37,7 @@ public class FileChannelService implements ChannelService {
         validate(channelDto);
         Channel ch = findById(id);
         ch.update(channelDto.getName(), channelDto.getDescription());
-        return channelRepository.save(ch);
+        return ch;
     }
 
     @Override
