--- src/main/java/com/sprint/mission/discodeit/service/jcf/JCFChannelService.java
@@ -1,65 +1,0 @@
-package com.sprint.mission.discodeit.service.jcf;
-
-import com.sprint.mission.discodeit.entity.Channel;
-import com.sprint.mission.discodeit.repository.ChannelRepository;
-import com.sprint.mission.discodeit.service.ChannelService;
-
-import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.UUID;
-
-public class JCFChannelService implements ChannelService {
-    private final ChannelRepository channelRepository;
-
-    public JCFChannelService(ChannelRepository channelRepository) {
-        this.channelRepository = channelRepository;
-    }
-
-    @Override
-    public Channel save(Channel channelDto) {
-        validate(channelDto);
-        return channelRepository.save(channelDto);
-    }
-
-    @Override
-    public Channel findById(UUID id) {
-        return channelRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("Channel not found: " + id));
-    }
-
-    @Override
-    public List<Channel> findAll() {
-        return channelRepository.findAll();
-    }
-
-    @Override
-    public Channel update(UUID id, Channel channelDto) {
-        validate(channelDto);
-        Channel ch = findById(id);
-        ch.update(channelDto.getName(), channelDto.getDescription());
-        return ch;
-    }
-
-    @Override
-    public void delete(UUID id) {
-        Channel ch = findById(id);
-        channelRepository.delete(ch.getId());
-    }
-
-    @Override
-    public void deleteAll() {
-        channelRepository.deleteAll();
-    }
-
-    private void validate(Channel chDto) {
-        if (chDto == null) {
-            throw new IllegalArgumentException("Channel must not be null");
-        }
-        if (chDto.getName() == null || chDto.getName().isBlank()) {
-            throw new IllegalArgumentException("Channel name is required");
-        }
-        if (chDto.getDescription() == null) {
-            throw new IllegalArgumentException("Channel description must not be null");
-        }
-    }
-}
